bot-project/requirements.txt
python-telegram-bot==20.7
openai==1.0.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.0
PyYAML==6.0.1
markdown==3.5.1
